requirements.txt
anyio==4.0.0
asgiref==3.7.1
boto3==1.28.22
botocore==1.31.22
Brotli==1.0.9
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.11
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.4.2
filelock==3.12.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
jmespath==1.0.1
msgpack==1.0.5
mysqlclient==2.2.0
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
proto-plus==1.22.3
protobuf==4.23.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.0
whitenoise==6.5.0
